requirements.txt
asgiref==3.5.1
autopep8==1.6.0
boto3==1.22.4
botocore==1.25.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
crispy-tailwind==0.5.0
cryptography==37.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.4
django-allauth==0.50.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-storages==1.12.3
django-utils-six==2.0
gunicorn==20.1.0
idna==3.3
jmespath==1.0.0
oauthlib==3.2.0
Pillow==9.1.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
